ml-service/requirements.txt
pandas>=1.5
numpy
scikit-learn
joblib
fastapi
uvicorn[standard]
python-dotenv
pydantic
